consolidate types
