+++ specification/watcher/activate_watch/ActivationStatus.ts
@@ -17,13 +17,12 @@
  * under the License.
  */
 
-import { Dictionary } from '@spec_utils/Dictionary'
-import { ActionStatus } from '@watcher/ack_watch/ActionStatus'
 import { ActivationState } from '@watcher/ack_watch/ActivationState'
-import { IndexName, VersionNumber } from '@_types/common'
+import { Actions } from '@watcher/_types/Action'
+import { VersionNumber } from '@_types/common'
 
 export class ActivationStatus {
-  actions: Dictionary<IndexName, ActionStatus>
+  actions: Actions
   state: ActivationState
   version: VersionNumber
 }
